adds rabbitmq messaging into the gateway.
Currently transitioning from a bad setup using fanout when it
should've been using direct.

Currently this is how rabbitmq works:

 * Gateway sends `v1.service.method`
 * A service responds to `v1.service.method`
 * A service does logic, responds with `v1.service.method.response`
 * Service sends response on `v1.staymarta.api` with request id routing key
 * Service response is routed to correct gateway via same method

Currently routing to the gateway that sent the response is *not* implemented, but it will be shortly (matter of uuid, or even using the docker container's ID)

A stress testing script has been included, which does 1000 requests to a basic *exisiting* service.

Signed-Off-By: Jared Allard <jaredallard@outlook.com>